aider: Added tests for traversal buttons triggering internal state changes.
# Aider chat conversation:

USER: Act as a world class software engineer with 30 years of experience. I want you to help me finish the tests for my visualisation traversal feature. The tests I already have are in `tests/e2e/HierarchyTraversalButtons.test.tsx`. Use the second describe block as an example. This shows how to setup and test a) the buttons are rendered when a given hierarchy mock is fed to the visualisation component, and also the coordinates that determine where in the visualisation we are (currentOrbitCoords, mocked with a jotai atom). E.g. when the hierarchy mock is called HIERARCHY_ROOT_TWO_CHILDREN_MOCKS I knnow that this is a balanced tree of three nodes. So it should trigger my 'incrementDepth' method when the down button is found and clicked. Also when I simulate being at coords of (x:0, y: 1) which would be the state after the down button is clicked, it should render the (traversal-button-up) button, as well as the (traversal-button-right) button, since these are the available options from that location. Use all of this information to first: 'when you click that button, it triggers internal currentOrbitCoords state to decrement y' and 'when you click that button, it triggers internal currentOrbitCoords state to increment x' and 'when you click that button, it triggers internal currentOrbitCoords state to decrement x', i.e. to test the other 3 callbacks. After that I will ask you to keep fleshing out the tests, but do this first before asking for the next ones.

ASSISTANT: ```diff
--- tests/e2e/HierarchyTraversalButtons.test.tsx
+++ tests/e2e/HierarchyTraversalButtons.test.tsx
@@ ... @@
 describe('Hierarchy Path Templates - it renders traversal buttons and triggers events - parent with 2 children', () => {
   let Tree;
   beforeEach(() => {
     Tree = renderVis(OrbitTree);
   });
   afterAll(() => {
     Tree = undefined;
   });

+  it('when you click the up button, it triggers internal currentOrbitCoords state to decrement y', async () => {
+    // Arrange
+    const { getByTestId } = render(
+      <MockedProvider mocks={HIERARCHY_ROOT_TWO_CHILDREN_MOCKS} addTypename={false}>
+        {WithCurrentOrbitCoordsMockedAtom(Tree, {x, y: 1})}</MockedProvider> );
+
+    // Act
+    await waitFor(() => {
+      expect(getByTestId('traversal-button-up')).toBeTruthy();
+    });
+    const traversalButton = getByTestId('traversal-button-up');
+    fireEvent.click(traversalButton);
+
+    // Assert
+    await waitFor(() => {
+      expect(decrementDepth).toHaveBeenCalled();
+    });
+  });
+
+  it('when you click the right button, it triggers internal currentOrbitCoords state to increment x', async () => {
+    // Arrange
+    const { getByTestId } = render(
+      <MockedProvider mocks={HIERARCHY_ROOT_TWO_CHILDREN_MOCKS} addTypename={false}>
+        {WithCurrentOrbitCoordsMockedAtom(Tree, {x: 0, y})}</MockedProvider> );
+
+    // Act
+    await waitFor(() => {
+      expect(getByTestId('traversal-button-right')).toBeTruthy();
+    });
+    const traversalButton = getByTestId('traversal-button-right');
+    fireEvent.click(traversalButton);
+
+    // Assert
+    await waitFor(() => {
+      expect(incrementBreadth).toHaveBeenCalled();
+    });
+  });
+
+  it('when you click the left button, it triggers internal currentOrbitCoords state to decrement x', async () => {
+    // Arrange
+    const { getByTestId } = render(
+      <MockedProvider mocks={HIERARCHY_ROOT_TWO_CHILDREN_MOCKS} addTypename={false}>
+        {WithCurrentOrbitCoordsMockedAtom(Tree, {x: 1, y})}</MockedProvider> );
+
+    // Act
+    await waitFor(() => {
+      expect(getByTestId('traversal-button-left')).toBeTruthy();
+    });
+    const traversalButton = getByTestId('traversal-button-left');
+    fireEvent.click(traversalButton);
+
+    // Assert
+    await waitFor(() => {
+      expect(decrementBreadth).toHaveBeenCalled();
+    });
+  });
 });
```